+++ Sources/Models/AppStoreAgeRating.swift
@@ -1,0 +1,15 @@
+//
+//  AppStoreAgeRating.swift
+//  AppStoreConnect-Swift-SDK
+//
+//  Created by Morten Bjerg Gregersen on 23/08/2020.
+//
+
+import Foundation
+
+public enum AppStoreAgeRating: String, Codable {
+    case fourPlus = "FOUR_PLUS"
+    case ninePlus = "NINE_PLUS"
+    case twelvePlus = "TWELVE_PLUS"
+    case seventeenPlus = "SEVENTEEN_PLUS"
+}
