+++ Sources/Models/AppStoreVersionState.swift
@@ -1,0 +1,28 @@
+//
+//  AppStoreVersionState.swift
+//  AppStoreConnect-Swift-SDK
+//
+//  Created by Morten Bjerg Gregersen on 23/08/2020.
+//
+
+import Foundation
+
+public enum AppStoreVersionState: String, Codable {
+    case developerRemovedFromSale = "DEVELOPER_REMOVED_FROM_SALE"
+    case developerRejected = "DEVELOPER_REJECTED"
+    case inReview = "IN_REVIEW"
+    case invalidBinary = "INVALID_BINARY"
+    case metadataRejected = "METADATA_REJECTED"
+    case pendingAppleRelease = "PENDING_APPLE_RELEASE"
+    case pendingContract = "PENDING_CONTRACT"
+    case pendingDeveloperRelease = "PENDING_DEVELOPER_RELEASE"
+    case prepareForSubmission = "PREPARE_FOR_SUBMISSION"
+    case preorderReadyForSale = "PREORDER_READY_FOR_SALE"
+    case processingForAppStore = "PROCESSING_FOR_APP_STORE"
+    case readyForSale = "READY_FOR_SALE"
+    case rejected = "REJECTED"
+    case removedFromSale = "REMOVED_FROM_SALE"
+    case waitingForExportCompliance = "WAITING_FOR_EXPORT_COMPLIANCE"
+    case waitingForReview = "WAITING_FOR_REVIEW"
+    case replacedWithNewVersion = "REPLACED_WITH_NEW_VERSION"
+}
