+++ Sources/Models/BrazilAgeRating.swift
@@ -1,0 +1,17 @@
+//
+//  BrazilAgeRating.swift
+//  AppStoreConnect-Swift-SDK
+//
+//  Created by Morten Bjerg Gregersen on 23/08/2020.
+//
+
+import Foundation
+
+public enum BrazilAgeRating: String, Codable {
+    case l = "L"
+    case ten = "TEN"
+    case twelve = "TWELVE"
+    case fourteen = "FOURTEEN"
+    case sixteen = "SIXTEEN"
+    case eighteen = "EIGHTEEN"
+}
