+++ Sources/Models/KidsAgeBand.swift
@@ -1,0 +1,14 @@
+//
+//  KidsAgeBand.swift
+//  AppStoreConnect-Swift-SDK
+//
+//  Created by Morten Bjerg Gregersen on 23/08/2020.
+//
+
+import Foundation
+
+public enum KidsAgeBand: String, Codable {
+    case fiveAndUnder = "FIVE_AND_UNDER"
+    case sixToEight = "SIX_TO_EIGHT"
+    case nineToEleven = "NINE_TO_ELEVEN"
+}
